+++ src/vapour_pressure_deficit.rs
@@ -5,6 +5,7 @@
 //!when it is saturated ([Wikipedia](https://en.wikipedia.org/wiki/Vapour-pressure_deficit)).
 
 use crate::{error_wrapper::InputError, vapour_pressure};
+use crate::Float;
 
 ///Formula for computing vapour pressure deficit from vapour pressure and saturation vapour pressure
 ///
@@ -13,7 +14,7 @@ use crate::{error_wrapper::InputError, vapour_pressure};
 ///Returns [`InputError::OutOfRange`] when one of inputs is out of range.\
 ///Valid `vapour_pressure` range: 0Pa - 10000Pa
 ///Valid `saturation_vapour_pressure` range: 0Pa - 10000Pa
-pub fn general1(vapour_pressure: f64, saturation_vapour_pressure: f64) -> Result<f64, InputError> {
+pub fn general1(vapour_pressure: Float, saturation_vapour_pressure: Float) -> Result<Float, InputError> {
     if !(0.0..=50_000.0).contains(&vapour_pressure) {
         return Err(InputError::OutOfRange(String::from("vapour_pressure")));
     }
@@ -35,7 +36,7 @@ pub fn general1(vapour_pressure: f64, saturation_vapour_pressure: f64) -> Result
 ///Returns [`InputError::OutOfRange`] when one of inputs is out of range.\
 ///Valid `vapour_pressure` range: 0Pa - 10000Pa
 ///Valid `saturation_vapour_pressure` range: 0Pa - 10000Pa
-pub fn general2(temperature: f64, dewpoint: f64, pressure: f64) -> Result<f64, InputError> {
+pub fn general2(temperature: Float, dewpoint: Float, pressure: Float) -> Result<Float, InputError> {
     if !(253.0..=324.0).contains(&temperature) {
         return Err(InputError::OutOfRange(String::from("temperature")));
     }
@@ -65,10 +66,10 @@ pub fn general2(temperature: f64, dewpoint: f64, pressure: f64) -> Result<f64, I
 ///Valid `vapour_pressure` range: 0Pa - 10000Pa
 ///Valid `saturation_vapour_pressure` range: 0Pa - 10000Pa
 pub fn general3(
-    temperature: f64,
-    relative_humidity: f64,
-    pressure: f64,
-) -> Result<f64, InputError> {
+    temperature: Float,
+    relative_humidity: Float,
+    pressure: Float,
+) -> Result<Float, InputError> {
     if !(253.0..=319.0).contains(&temperature) {
         return Err(InputError::OutOfRange(String::from("temperature")));
     }
